chore(structure): rename features folder to modules
