+++ backend/nestjs/src/app.module.ts
@@ -1,12 +1,12 @@
 import { Module } from "@nestjs/common";
 import { ConfigModule } from "./config/config.module";
 import { DrizzleModule } from "./drizzle/drizzle.module";
-import { AuthModule } from "./features/auth/auth.module";
+import { AuthModule } from "./modules/auth/auth.module";
 import { RepositoryModule } from "./repository/repository.module";
-import { EmailModule } from "./features/email/email.module";
-import { ProductModule } from "./features/product/product.module";
+import { EmailModule } from "./modules/email/email.module";
+import { ProductModule } from "./modules/product/product.module";
 import { ScheduleModule } from "@nestjs/schedule";
-import { ShopModule } from "./features/shop/shop.module";
+import { ShopModule } from "./modules/shop/shop.module";
 
 @Module({
   imports: [
