+++ backend/nestjs/src/modules/email/interfaces/service.ts
