+++ backend/nestjs/src/modules/shop/interfaces/service.ts
