+++ backend/nestjs/src/modules/shop/shop.service.ts
